[22.11.16 성태웅] 삭제기능 추가 + script 위치 조절

+++ src/main/resources/static/js/board/content.js
@@ -6,18 +6,22 @@ const $delBtn = document.getElementById('del-btn');
 
 //목록버튼
 $listBtn.onclick = e => {
-    confirm('목록가기?');
+
     location.href = '/board?pageNum=${p.pageNum}&amount=${p.amount}';
 };
 
 //수정버튼
 $editBtn.onclick = e => {
-    confirm('수정하시겠습니까?');
-    location.href='/board/edit?boardNo=${b.boardNo}';
+    if($editBtn!==null) {
+        confirm('수정하시겠습니까?');
+        location.href='/board/edit?boardNo=${b.boardNo}';
+    }
 };
 
 //삭제버튼
 $delBtn.onclick = e => {
-    confirm('삭제하시겠습니까?');
-    location.href='/board/delete?boardNo=${b.boardNo}';
+    if($delBtn!==null) {
+        confirm('삭제하시겠습니까?');
+        location.href='/board/delete?boardNo=${b.boardNo}';
+    }
 };
